Refactors the code.
